refactor(player): merge native.rs into mod.rs
The native.rs module was a thin indirection — its contents (PlaybackResult,
play_session, play_session_native, run_main_loop) belong directly in the
player module root. This removes the extra file and re-export layer while
keeping the combined file under the 400-line limit (372 lines).
